Created node classes
The purpose of these classes is to allow for manipulating the query
AST.
To do this they store the relevant information about a node, as well
as a reference to it.
Neither the reference to the original node, nor the contests of it,
should be changed. This information is intended to be used to locate
the node in the AST.

The details stored in the wrapper classes can however be changed, and
used to create a new node.

The old node can be located using the reference to it, and it can then
be replaced with a new node with updated information.