Refactored QueryStructure to use TableNode and ColumnNode
Also improved the handling of subqueries because I broke it.
The issue was that the same table occured in a query and its subquery.
Because the table was repeated in the subquery it was given an alias,
even though it was not needed, because the subquery is in a new scope.

It has been changed such that before a query is transformed, all
subqueries are removed from the query and transformed first.
Then the query itself is transformed, and the transformed subqueries
are replaced.

The transformation of subqueries should be entirely independant, and
it should be possible to transform them in parallel, but that just
seemed like a lot of work.